create utils submodule and refactor
