+++ gdsfactory/simulation/gmeep/write_sparameters_meep_mpi.py
@@ -200,7 +200,7 @@ def write_sparameters_meep_mpi(
     if live_output:
         import asyncio
 
-        from gdsfactory.async_utils import execute_and_stream_output
+        from gdsfactory.utils.async_utils import execute_and_stream_output
 
         asyncio.run(
             execute_and_stream_output(
